Fixed role change request
